@@ -1,93 +1,107 @@
 {
-    "name": "@scaffold-eth/react-app",
-    "version": "1.1.0",
-    "homepage": ".",
-    "browserslist": {
-        "production": [
-            ">0.2% or last 2 versions",
-            "not dead"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    },
-    "dependencies": {
-        "@davatar/react": "^1.8.1",
-        "@emotion/react": "^11.5.0",
-        "@emotion/styled": "^11.3.0",
-        "@ethersproject/providers": "^5.0.12",
-        "@material-ui/core": "^4.11.4",
-        "@material-ui/icons": "^4.11.2",
-        "@material-ui/lab": "^4.0.0-alpha.58",
-        "@mui/material": "^5.1.0",
-        "@netlify/functions": "^0.8.0",
-        "@reduxjs/toolkit": "^1.6.1",
-        "@walletconnect/web3-provider": "^1.6.5",
-        "axios": "^0.21.2",
-        "classnames": "^2.3.1",
-        "ethers": "^5.4.0",
-        "from-exponential": "^1.1.1",
-        "i18next": "^21.5.4",
-        "i18next-browser-languagedetector": "^6.1.2",
-        "intl": "^1.2.5",
-        "lodash": "^4.17.21",
-        "node-sass": "^6.0.1",
-        "node-watch": "^0.7.1",
-        "notistack": "^1.0.10",
-        "numeral": "^2.0.6",
-        "react": "^16.14.0",
-        "react-copy-to-clipboard": "^5.0.4",
-        "react-dom": "^16.14.0",
-        "react-i18next": "^11.14.3",
-        "react-redux": "^7.2.4",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "^4.0.3",
-        "react-snowflakes": "^1.0.0",
-        "react-snowstorm": "^0.0.3",
-        "redux": "^4.1.0",
-        "redux-logger": "^3.0.6",
-        "redux-thunk": "^2.3.0",
-        "sass": "^1.35.1",
-        "typescript": "^4.3.2",
-        "web3modal": "^1.9.4"
-    },
-    "devDependencies": {
-        "@types/lodash": "^4.14.172",
-        "@types/node": "^14.17.32",
-        "@types/numeral": "^2.0.2",
-        "@types/react": "^17.0.15",
-        "@types/react-copy-to-clipboard": "^5.0.2",
-        "@types/react-dom": "^17.0.9",
-        "@types/react-router-dom": "^5.1.7",
-        "@typescript-eslint/eslint-plugin": "^4.28.5",
-        "@typescript-eslint/parser": "^4.28.5",
-        "babel-plugin-module-resolver": "^4.1.0",
-        "eslint": "^7.31.0",
-        "eslint-config-airbnb": "^18.2.0",
-        "eslint-config-prettier": "^6.11.0",
-        "eslint-plugin-babel": "^5.3.1",
-        "eslint-plugin-import": "^2.22.1",
-        "eslint-plugin-jsx-a11y": "^6.4.1",
-        "eslint-plugin-prettier": "^3.1.4",
-        "eslint-plugin-react": "^7.22.0",
-        "eslint-plugin-react-hooks": "^4.2.0",
-        "prettier": "^2.0.5"
-    },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "resolutions": {
-        "react-error-overlay": "6.0.9"
-    },
-    "scripts": {
-        "build": "GENERATE_SOURCEMAP=false react-scripts build",
-        "eject": "react-scripts eject",
-        "start": "react-scripts start",
-        "debug": "BROWSER='google chrome' BROWSER_ARGS='--remote-debugging-port=9222' react-scripts start",
-        "test": "react-scripts test",
-        "lint": "npx eslint src/ --ext .ts,.js,.tsx,.jsx",
-        "lint-fix": "npx eslint src/ --ext .ts,.js,.tsx,.jsx --fix"
-    }
+  "name": "@scaffold-eth/react-app",
+  "version": "1.1.0",
+  "homepage": ".",
+  "browserslist": {
+    "production": [
+      ">0.2% or last 2 versions",
+      "not dead"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "dependencies": {
+    "@davatar/react": "^1.8.1",
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
+    "@ethersproject/providers": "^5.0.12",
+    "@material-ui/core": "^4.11.4",
+    "@material-ui/icons": "^4.11.2",
+    "@material-ui/lab": "^4.0.0-alpha.58",
+    "@mui/icons-material": "^5.5.1",
+    "@mui/material": "^5.1.0",
+    "@netlify/functions": "^0.8.0",
+    "@reduxjs/toolkit": "^1.6.1",
+    "@walletconnect/web3-provider": "^1.6.5",
+    "axios": "^0.21.2",
+    "classnames": "^2.3.1",
+    "ethers": "^5.4.0",
+    "from-exponential": "^1.1.1",
+    "i18next": "^21.5.4",
+    "i18next-browser-languagedetector": "^6.1.2",
+    "intl": "^1.2.5",
+    "lodash": "^4.17.21",
+    "node-sass": "^6.0.1",
+    "node-watch": "^0.7.1",
+    "notistack": "^1.0.10",
+    "numeral": "^2.0.6",
+    "react": "^16.14.0",
+    "react-copy-to-clipboard": "^5.0.4",
+    "react-dom": "^16.14.0",
+    "react-i18next": "^11.14.3",
+    "react-redux": "^7.2.4",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^4.0.3",
+    "react-snowflakes": "^1.0.0",
+    "react-snowstorm": "^0.0.3",
+    "redux": "^4.1.0",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0",
+    "sass": "^1.35.1",
+    "typescript": "^4.3.2",
+    "web3modal": "^1.9.4"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.17.5",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@nomiclabs/hardhat-ethers": "^2.0.5",
+    "@nomiclabs/hardhat-etherscan": "^3.0.3",
+    "@nomiclabs/hardhat-waffle": "^2.0.3",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.3",
+    "@typechain/hardhat": "^5.0.0",
+    "@types/chai": "^4.3.0",
+    "@types/lodash": "^4.14.172",
+    "@types/node": "^14.18.12",
+    "@types/numeral": "^2.0.2",
+    "@types/react": "^17.0.15",
+    "@types/react-copy-to-clipboard": "^5.0.2",
+    "@types/react-dom": "^17.0.9",
+    "@types/react-router-dom": "^5.1.7",
+    "@typescript-eslint/eslint-plugin": "^4.28.5",
+    "@typescript-eslint/parser": "^4.28.5",
+    "cross-env": "^7.0.3",
+    "eslint": "^7.31.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-babel": "^5.3.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "ethers": "^5.0.0",
+    "hardhat": "^2.9.1",
+    "hardhat-contract-sizer": "^2.5.0",
+    "hardhat-gas-reporter": "^1.0.8",
+    "prettier": "^2.0.5",
+    "solidity-coverage": "^0.7.20",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.2"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "scripts": {
+    "build": "react-scripts build",
+    "eject": "react-scripts eject",
+    "start": "react-scripts start",
+    "test": "react-scripts test",
+    "lint": "npx eslint src/ --ext .ts,.js,.tsx,.jsx",
+    "lint-fix": "npx eslint src/ --ext .ts,.js,.tsx,.jsx --fix"
+  }
 }
