@@ -1,20 +1,20 @@
 import { useEffect, useState, useCallback } from "react";
 import { Route, Redirect, Switch } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { useAddress, useWeb3Context } from "../hooks";
-import { calcBondDetails } from "../store/slices/bond-slice";
-import { loadAppDetails } from "../store/slices/app-slice";
-import { loadAccountDetails, calculateUserBondDetails, calculateUserTokenDetails } from "../store/slices/account-slice";
-import { IReduxState } from "../store/slices/state.interface";
-import Loading from "../components/Loader";
-import useBonds from "../hooks/bonds";
-import ViewBase from "../components/ViewBase";
+import { useAddress, useWeb3Context } from "hooks/web3";
+import { calcBondDetails } from "store/slices/bond-slice";
+import { loadAppDetails } from "store/slices/app-slice";
+import { loadAccountDetails, calculateUserBondDetails, calculateUserTokenDetails } from "store/slices/account-slice";
+import { IReduxState } from "store/slices/state.interface";
+import Loading from "components/Loader";
+import useBonds from "hooks/bonds";
+import ViewBase from "layout/ViewBase";
 import { Stake, Forecast, ChooseBond, Bond, Dashboard, NotFound, Redeem, Wrap } from "../views";
 
 import "./style.scss";
 import useTokens from "../hooks/tokens";
+import { BaseLayout } from "layout/Base";
 import { Typography } from "@material-ui/core";
-import { WalletNotConnected } from "components/WalletNotConnected";
 
 function App() {
     const dispatch = useDispatch();
@@ -29,8 +29,8 @@ function App() {
     });
     const isAppLoaded = useSelector<IReduxState, boolean>(state => !Boolean(state.app.marketPrice));
 
-    // const { bonds } = useBonds();
-    // const { tokens } = useTokens();
+    const { bonds } = useBonds();
+    const { tokens } = useTokens();
 
     async function loadDetails(whichDetails: string) {
         let loadProvider = provider;
@@ -39,35 +39,35 @@ function App() {
             loadApp(loadProvider);
         }
 
-        // if (whichDetails === "account" && address && connected) {
-        //     loadAccount(loadProvider);
-        //     if (isAppLoaded) return;
+        if (whichDetails === "account" && address && connected) {
+            loadAccount(loadProvider);
+            if (isAppLoaded) return;
 
-        //     // loadApp(loadProvider);
-        // }
+            loadApp(loadProvider);
+        }
 
-        // if (whichDetails === "userBonds" && address && connected) {
-        //     bonds.map(bond => {
-        //         dispatch(calculateUserBondDetails({ address, bond, provider, networkID: chainID }));
-        //     });
-        // }
+        if (whichDetails === "userBonds" && address && connected) {
+            bonds.map(bond => {
+                dispatch(calculateUserBondDetails({ address, bond, provider, networkID: chainID }));
+            });
+        }
 
-        // if (whichDetails === "userTokens" && address && connected) {
-        //     tokens.map(token => {
-        //         dispatch(calculateUserTokenDetails({ address, token, provider, networkID: chainID }));
-        //     });
-        // }
+        if (whichDetails === "userTokens" && address && connected) {
+            tokens.map(token => {
+                dispatch(calculateUserTokenDetails({ address, token, provider, networkID: chainID }));
+            });
+        }
     }
 
     const loadApp = useCallback(
         loadProvider => {
             dispatch(loadAppDetails({ networkID: chainID, provider: loadProvider }));
-            // bonds.map(bond => {
-            //     dispatch(calcBondDetails({ bond, value: null, provider: loadProvider, networkID: chainID }));
-            // });
-            // tokens.map(token => {
-            //     dispatch(calculateUserTokenDetails({ address: "", token, provider, networkID: chainID }));
-            // });
+            bonds.map(bond => {
+                dispatch(calcBondDetails({ bond, value: null, provider: loadProvider, networkID: chainID }));
+            });
+            tokens.map(token => {
+                dispatch(calculateUserTokenDetails({ address: "", token, provider, networkID: chainID }));
+            });
         },
         [connected],
     );
@@ -103,12 +103,8 @@ function App() {
     return (
         <ViewBase>
             <Switch>
-                <Route path="/">
-                    <Dashboard />
-                </Route>
-                {/* 
                 <Route exact path="/">
-                    <Redirect to="/stake" />
+                    <Dashboard />
                 </Route>
 
                 <Route path="/stake">
@@ -125,9 +121,9 @@ function App() {
 
                 <Route path="/Forecast">
                     <Forecast />
-                </Route> */}
+                </Route>
 
-                {/* <Route path="/mints">
+                <Route path="/mints">
                     {bonds.map(bond => {
                         return (
                             <Route exact key={bond.name} path={`/mints/${bond.name}`}>
@@ -136,7 +132,7 @@ function App() {
                         );
                     })}
                     <ChooseBond />
-                </Route> */}
+                </Route>
 
                 <Route component={NotFound} />
             </Switch>
