@@ -1,9 +1,9 @@
 import { SvgIcon, Link } from "@material-ui/core";
-import { ReactComponent as GitHub } from "../../../assets/icons/github.svg";
-import { ReactComponent as Twitter } from "../../../assets/icons/twitter.svg";
-import { ReactComponent as Telegram } from "../../../assets/icons/telegram.svg";
-import { ReactComponent as Discord } from "../../../assets/icons/discord.svg";
-import DocsIcon from "../../../assets/icons/docs.svg";
+import { ReactComponent as GitHub } from "assets/icons/github.svg";
+import { ReactComponent as Twitter } from "assets/icons/twitter.svg";
+import { ReactComponent as Telegram } from "assets/icons/telegram.svg";
+import { ReactComponent as Discord } from "assets/icons/discord.svg";
+import DocsIcon from "assets/icons/docs.svg";
 
 export default function Social() {
     return (
